# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GAB SCHOOL CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: trainingType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: groupId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Register a new student (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /students/{id}/stage:
    patch:
      operationId: updateStudentStage
      tags: [students]
      summary: Change student pipeline stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageBody"
      responses:
        "200":
          description: Stage updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
  /students/{id}/group:
    patch:
      operationId: assignStudentToGroup
      tags: [students]
      summary: Assign student to a group
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignGroupBody"
      responses:
        "200":
          description: Group assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
...
components:
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        trainingType:
          type: string
          enum: [online, physical]
        housingNeeded:
          type: boolean
        experienceLevel:
          type: string
        note:
          type: ["string", "null"]
        stage:
          type: string
          enum: [new, contacted, interested, no_show, archived]
        groupId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - phone
        - whatsapp
        - city
        - trainingType
        - housingNeeded
        - experienceLevel
        - stage
        - createdAt
        - updatedAt
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        trainingType:
          type: string
          enum: [online, physical]
        housingNeeded:
          type: boolean
        experienceLevel:
          type: string
        note:
          type: string
      required:
        - firstName
        - lastName
        - phone
        - whatsapp
        - city
        - trainingType
        - housingNeeded
        - experienceLevel
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        trainingType:
          type: string
          enum: [online, physical]
        housingNeeded:
          type: boolean
        experienceLevel:
          type: string
        note:
          type: ["string", "null"]
        stage:
          type: string
          enum: [new, contacted, interested, no_show, archived]
    UpdateStageBody:
      type: object
      properties:
        stage:
          type: string
          enum: [new, contacted, interested, no_show, archived]
      required:
        - stage
    AssignGroupBody:
      type: object
      properties:
        groupId:
          type: ["integer", "null"]
      required:
        - groupId
...