# lib/api-zod/src/generated/types/updateStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAB SCHOOL CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStudentBodyStage } from "./updateStudentBodyStage";
import type { UpdateStudentBodyTrainingType } from "./updateStudentBodyTrainingType";

export interface UpdateStudentBody {
  firstName?: string;
  lastName?: string;
  phone?: string;
  whatsapp?: string;
  city?: string;
  trainingType?: UpdateStudentBodyTrainingType;
  housingNeeded?: boolean;
  experienceLevel?: string;
  /** @nullable */
  note?: string | null;
  stage?: UpdateStudentBodyStage;
}